libmpdclient: use size_t and ssize_t
When working with buffer lengths and recv() results, we should declare
size_t and ssize_t variables, otherwise the code may fail on 64 bit
machines.

+++ src/libmpdclient.h
@@ -40,7 +40,8 @@
 #endif
 
 #include <sys/time.h>
-#include <stdarg.h>
+#include <stddef.h>
+
 #define MPD_BUFFER_MAX_LENGTH	50000
 #define MPD_WELCOME_MESSAGE	"OK MPD "
 
@@ -117,8 +118,8 @@ typedef struct _mpd_Connection {
 	/* DON'T TOUCH any of the rest of this stuff */
 	int sock;
 	char buffer[MPD_BUFFER_MAX_LENGTH+1];
-	int buflen;
-	int bufstart;
+	size_t buflen;
+	size_t bufstart;
 	int doneProcessing;
 	int listOks;
 	int doneListOk;
